requirements.txt
absl-py==0.9.0
astunparse==1.6.3
attrs==19.3.0
Babel==2.8.0
bz2file==0.98
cachetools==4.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
decorator==4.4.2
dill==0.3.2
dopamine-rl==3.0.1
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gast==0.3.3
gevent==20.6.2
gin-config==0.3.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.30.0
gunicorn==20.0.4
gym==0.17.2
h5py==2.10.0
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
Keras-Preprocessing==1.1.2
kfac==0.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
mesh-tensorflow==0.1.13
mpmath==1.1.0
nltk==3.5
numpy==1.19.0
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.4
pandas==1.0.5
Pillow==7.1.2
pkg-resources==0.0.0
portalocker==1.7.0
promise==2.3
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
pyparsing==2.4.7
pypng==0.0.20
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.6
sacrebleu==1.4.10
sacremoses==0.0.43
scikit-learn==0.23.1
scipy==1.4.1
sentencepiece==0.1.91
six==1.15.0
sympy==1.6
t5==0.6.0
tensor2tensor==1.15.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-addons==0.10.0
tensorflow-datasets==3.1.0
tensorflow-estimator==2.2.0
tensorflow-gan==2.0.0
tensorflow-hub==0.8.0
tensorflow-metadata==0.22.2
tensorflow-probability==0.7.0
tensorflow-text==2.2.1
termcolor==1.1.0
tf-slim==1.1.0
tfds-nightly==3.1.0.dev202006230105
threadpoolctl==2.1.0
tokenizers==0.7.0
torch==1.5.1
tqdm==4.46.1
transformers==2.11.0
typeguard==2.9.1
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
